requirements.txt
asgiref==3.5.0
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.1
django-cors-headers==3.13.0
django-heroku==0.3.1
djangorestframework==3.13.1
filelock==3.4.2
gunicorn==20.1.0
idna==3.3
Pillow==9.0.0
platformdirs==2.4.1
psycopg2==2.9.3
pytz==2022.1
qrcode==7.3.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.10
virtualenv==20.13.0
whitenoise==6.0.0
